+++ CHANGELOG.md
@@ -5,6 +5,42 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added `get_config` method to `request::Client` to fetch the configuration from
+  the Manager server.
+
+### Changed
+
+- Replaced `reload_config` request handler with `update_config`.
+  - The `reload_config` relied on a local temporary file to change the configuration.
+  - The `update_config` now results in fetching the configuration from the Manager
+    server.
+- Two Crusher modes introduced by `update_config`:
+  - Remote mode (no configuration file is provided using the `-c` CLI option):
+    - Attempts to fetch the configuration from the Manager server on startup.
+    - Re-fetches and applies the remote configuration upon receiving an
+      `update_config` request from the Manager server.
+    - If a panic occurs, enters a wait mode to await an `update_config` request
+      from the Manager server.
+    - Logs are printed to stdout while in wait mode.
+  - Local mode (a configuration file is provided using the `-c` CLI option):
+    - Ignores `update_config` requests from the Manager server.
+    - Does not enter wait mode, even in case of a panic.
+  - Common behaviors:
+    - Logs are printed to stdout if `log_dir` is not specified in the
+      configuration.
+    - Panics if the CLI arguments for the Manager server are incorrect.
+- Changed `request::Client` structure to support `get_config` method.
+- Moved `to_cert_chain`, `to_private_key`, `to_ca_certs` functions to `Certs`'s
+  methods for modularity.
+
+### Removed
+
+- Removed default configuration file: /usr/local/aice/conf/crusher.toml
+
 ## [0.5.0] - 2024-11-26
 
 ### Changed
@@ -135,6 +171,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Send the generated `time series` to `giganto`'s ingest.
 - Save the model's id and the last time the timeseries was sent to a file.
 
+[Unreleased]: https://github.com/aicers/crusher/compare/0.5.0...main
 [0.5.0]: https://github.com/aicers/crusher/compare/0.4.1...0.5.0
 [0.4.1]: https://github.com/aicers/crusher/compare/0.4.0...0.4.1
 [0.4.0]: https://github.com/aicers/crusher/compare/0.3.2...0.4.0